fix:modify types in the database migration

+++ migrations/sqls/20211018141355-orders-up.sql
@@ -1,1 +1,1 @@
-CREATE TABLE orders(id SERIAL PRIMARY KEY,status VARCHAR(64),user_id bigint REFERENCES users(id));
+CREATE TABLE orders(id SERIAL PRIMARY KEY,status VARCHAR(64),user_id integer REFERENCES users(id));
